整理代码

+++ include/kernel/cpu.h
@@ -90,7 +90,7 @@ struct GATE_DESCRIPTOR {
 };
 
 void set_segmdesc(struct SEGMENT_DESCRIPTOR *sd, u32 limit, int base, int ar);
-void set_gatedesc(struct GATE_DESCRIPTOR *gd, int offset, int selector, int ar);
+void set_gatedesc(struct GATE_DESCRIPTOR *gd, size_t offset, int selector, int ar);
 
 void asm_error0();
 void asm_error1();
