+++ include/loader/cpu.h
@@ -31,4 +31,4 @@ struct GATE_DESCRIPTOR {
 };
 
 void init_gdtidt(void);
-void set_gatedesc(struct GATE_DESCRIPTOR *gd, int offset, int selector, int ar);
+void set_gatedesc(struct GATE_DESCRIPTOR *gd, size_t offset, int selector, int ar);
